refactor: update import paths for tags management
