--- src/features/tagManagement/api/tagManager.ts
@@ -1,7 +1,0 @@
-import { createTag } from './createTag';
-import { deleteTag } from './deleteTag';
-import { fetchTag } from './fetchTag';
-import { fetchTags } from './fetchTags';
-import { updateTag } from './updateTag';
-
-export { createTag, deleteTag, fetchTag, fetchTags, updateTag };
